Some changes for compilation with xlc
git-svn-id: file:///home/roberto/Git/svn@2651 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: fafad50dfac29f12fc136ccc0cd144752c1c654c

+++ trunk/xmipp/external/bilib/sources/waveletfiltersfract.cc
@@ -65,14 +65,14 @@ extern int WaveletFiltersGetSize_Fract(double Alpha, long *nh, long *ng)
 
 extern int WaveletFiltersGetCoef_Fract(double Alpha, double *h, double *g) 
 {
-static double *R, *I;
-double dzeta,mulfact;
-int i,n,m,sumTerms,size;
-unsigned long t;
+    static double *R, *I;
+    double dzeta,mulfact;
+    int i,n,m,sumTerms,size;
+    unsigned long t;
 
-	Alpha=fabs(Alpha);
-        size=getFilterSize(Alpha);
-	m=(int)ceil(log((long double)size)/log(2.0))+1;
+    Alpha=fabs(Alpha);
+    size=getFilterSize(Alpha);
+    m=(int)ceil(log((double)size)/log(2.0))+1;
     n=1<<m;
     while(2*n<size)
     { /* So that we're sure we have enough points ! */
